requirements.txt
python-telegram-bot
ramapi
requests